pyproject.toml
[tool.poetry]
name = "web-taskswift-django"
version = "0.1.0"
description = ""
authors = ["Roman <anovelforstudy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
icecream = "^2.1.3"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-jet-reboot = "^1.3.7"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}
ipython = "^8.22.1"
pillow = "^10.2.0"
django-colorfield = "^0.11.0"
djlint = "^1.34.1"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
djoser = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 119

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.isort]
known-third-party = ["icecream"]
known-first-party = ["src", "api", "users", "tasks", "notes",]
